+++ internal/pcisysfs/config/types.go
